requirements.txt
# Core Service Framework
fastapi
uvicorn[standard]

# LLM & Async HTTP
openai
httpx
aiohttp

# Parsing & Data
python-libnmap
sqlalchemy
psycopg2-binary
python-dotenv

# Security & Auth
python-jose[cryptography]
passlib[bcrypt]
python-keycloak

# Task Queue
rq

# ML & Testing
scikit-learn
pytest
locust
uv
